requirements.txt
anthropic==0.42.0
databases==0.9.0
fastapi==0.115.6
greenlet==3.1.1
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
SQLAlchemy==2.0.36
sqlite-fts4==1.0.3
sqlite-utils==3.38
typing_extensions==4.12.2
uvicorn==0.34.0